move dialog components to Dialog subfolder
