+++ client/src/pages/Dashboard/Dialogs/ViewCoffeeDialog.jsx
