+++ client/src/pages/Dashboard/index.js
@@ -1,6 +1,6 @@
 export { default as CoffeeCard } from './CoffeeCard';
 export { default as CoffeeCardMenu } from './CoffeeCardMenu';
-export { default as AddCoffeeDialog } from './AddCoffeeDialog';
-export { default as ViewCoffeeDialog } from './ViewCoffeeDialog';
-export { default as EditCoffeeDialog } from './EditCoffeeDialog';
-export { default as DataViewField } from './DataViewField';
+export { default as AddCoffeeDialog } from './Dialogs/AddCoffeeDialog';
+export { default as ViewCoffeeDialog } from './Dialogs/ViewCoffeeDialog';
+export { default as EditCoffeeDialog } from './Dialogs/EditCoffeeDialog';
+export { default as DataViewField } from './Dialogs/DataViewField';
